server/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiohttp-retry==2.9.1
aiosignal==1.3.2
alembic==1.16.1
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.3
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.2.1
colorama==0.4.6
Flask==3.1.1
flask-cors==6.0.0
flask-marshmallow==1.3.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
frozenlist==1.5.0
greenlet==3.2.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==5.3.1
Mako==1.3.10
MarkupSafe==3.0.2
marshmallow==4.0.0
marshmallow-sqlalchemy==1.4.2
multidict==6.1.0
numpy==2.2.2
pandas==2.2.3
propcache==0.3.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.1
requests==2.32.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.41
twilio==9.5.0
typing_extensions==4.13.2
tzdata==2025.1
urllib3==2.3.0
websocket-client==1.8.0
websockets==14.2
Werkzeug==3.1.3
yarl==1.18.3
